requirements.txt
pytelegrambotapi==4.26.0
playwright==1.52.0
requests==2.32.3
python-dotenv==1.0.1
paddleocr==2.8.1
paddlepaddle==3.3.1
google-cloud-documentai==2.30.0
